pyproject.toml
[project]
name = "plate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.8.1",
    "diffusers>=0.34.0",
    "matplotlib>=3.10.3",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.0",
    "opencv-python>=4.11.0.86",
    "pycocotools>=2.0.10",
    "segment-anything",
    "sentencepiece>=0.2.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.1",
    "websocket-client>=1.8.0",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git" }
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
